+++ mathstats/mean/rolling_mean.py
@@ -46,7 +46,7 @@ class RollingMean(simple_mean.SimpleMean):
         )
 
     @property
-    def gmean(self) -> np.array:
+    def gmean(self) -> np.ndarray:
         """
         Returns
         -------
@@ -68,7 +68,7 @@ class RollingMean(simple_mean.SimpleMean):
         )
 
     @property
-    def windowed_outgoing_row(self) -> np.array:
+    def windowed_outgoing_row(self) -> np.ndarray:
         """
         Return the outgoing row (FIFO - first in first out)
         of the rolling window
@@ -83,7 +83,7 @@ class RollingMean(simple_mean.SimpleMean):
     def update(
         self,
         incoming_variables: Union[np.ndarray, pd.Series],
-        outgoing_variables: np.array,
+        outgoing_variables: np.ndarray,
         **kwargs,
     ) -> None:
         """