+++ visualization/pdf_creator/plots/line_chart.py
@@ -25,8 +25,8 @@ class LineChart(object):
 
     def __init__(
         self,
-        x: Union[list, np.array, pd.Series],
-        y: Union[list, np.array, pd.Series],
+        x: Union[list, np.ndarray, pd.Series],
+        y: Union[list, np.ndarray, pd.Series],
         title: str = "",
         height: int = 350,
         width: int = None,